Impementing s flag handler into get_immunization_by_id
